fix type conversion

+++ tests/datawizard/gpu_register.c
@@ -65,12 +65,12 @@ find_a_worker(enum starpu_archtype type)
 static int
 check_result(unsigned *t, size_t size)
 {
-	int i;
-	for (i = 0; i < (int)size; i++)
+	unsigned i;
+	for (i = 0; i < size; i++)
 	{
 		if (t[i] != i*2)
 		{
-			FPRINTF(stderr,"t[%d] is %u instead of %d\n", i, t[i], 2*i);
+			FPRINTF(stderr,"t[%d] is %u instead of %u\n", i, t[i], 2*i);
 			return 1;
 		}
 	}
