+++ tests/datawizard/handle_to_pointer.c
@@ -25,7 +25,7 @@ static void cpu_task(void **buffers, void *args)
 {
 	int *numbers;
 	int i;
-	size_t size;
+	int size;
 
 	numbers = (int *) STARPU_VECTOR_GET_PTR(buffers[0]);
 	starpu_codelet_unpack_args (args, &size);
@@ -41,7 +41,7 @@ static void cuda_task(void **buffers, void *args)
 {
 	int *numbers;
 	int i;
-	size_t size;
+	int size;
 
 	numbers = (int *)STARPU_VECTOR_GET_PTR(buffers[0]);
 	starpu_codelet_unpack_args (args, &size);
@@ -98,10 +98,10 @@ static struct starpu_codelet cl =
 int main(int argc, char *argv[])
 {
 	int err, ret;
-	size_t i;
+	int i;
 	int *pointer;
 	starpu_data_handle_t handle;
-	static const size_t count = 123;
+	static const int count = 123;
 
 	ret = starpu_init(NULL);
 	if (ret == -ENODEV) return STARPU_TEST_SKIPPED;
